Fix for workitem 6279:  "extract All" on WinXP produces a spurious message that an empty directory in the zip archive is password protected.  Turns out that if you append a trailing slash to the directory name, you don't get this behavior.  Updated the ZipEntry.cs module to do this for directories. Also updated a few tests that specifically looked at directory names.
This introduces a change in the behavior of DotNetZip- now, all directory names inserted into an archive will have a trailing slash.  This should be totally compatible with other zip tools. 

totally unrelated change: I also changed the name on the output assembly for the winforms example to be DotNetZip-WinformsTool. 

git-svn-id: https://dotnetzip.svn.codeplex.com/svn@25202 4ef1bfeb-0a34-4d02-9305-240e5b07a168