Another check-in for workitem 6437 - ZIP64 support.  With this change set, the library can read or write ZIP64 archives.  I believe the library is consistent with itself - it will be able to read any archive it generates.  But, I am not clear on how interoperable this library will be with other ZIP64-aware tools.
Also included in this checkin are myriad doc updates, a new module for tests for ZIP64, an update to the Zipit command line tool to do ZIP64, and an update to the WinForms GUI tool to do ZIP64.

In addition, I removed some older properties on ZipFile - like TrimFullyQualifiedPaths.  I don't think anyone is using them. 


git-svn-id: https://dotnetzip.svn.codeplex.com/svn@27076 4ef1bfeb-0a34-4d02-9305-240e5b07a168